Moved parse function out of block, added tests
