+++ src/oauth_helpers.js
@@ -104,3 +104,32 @@ export async function refreshAuthTokens(refreshToken) {
     return items;
   }
 }
+
+export function fetchQueryParams(str) {
+  if (typeof str !== "string") {
+    return {};
+  }
+  str = str.trim().replace(/^(\?|#|&)/, "");
+  if (!str) {
+    return {};
+  }
+  return str.split("&").reduce(function (ret, param) {
+    var parts = param.replace(/\+/g, " ").split("=");
+    // Firefox (pre 40) decodes `%3D` to `=`
+    // https://github.com/sindresorhus/query-string/pull/37
+    var key = parts.shift();
+    var val = parts.length > 0 ? parts.join("=") : undefined;
+    key = decodeURIComponent(key);
+    // missing `=` should be `null`:
+    // http://w3.org/TR/2012/WD-url-20120524/#collect-url-parameters
+    val = val === undefined ? null : decodeURIComponent(val);
+    if (!ret.hasOwnProperty(key)) {
+      ret[key] = val;
+    } else if (Array.isArray(ret[key])) {
+      ret[key].push(val);
+    } else {
+      ret[key] = [ret[key], val];
+    }
+    return ret;
+  }, {});
+}
