+++ src/oauth_helpers.test.js
@@ -1,6 +1,7 @@
 import { ORBIT_HEADERS } from "./constants";
 import {
   configureRequest,
+  fetchQueryParams,
   isOAuthTokenExpired,
   refreshAuthTokens,
 } from "./oauth_helpers";
@@ -153,4 +154,19 @@ test("refreshAuthTokens unsets tokens if the request fails, for example if the r
   });
 
   global.chrome = originalChrome;
+});
+
+test("fetchQueryParams returns empty object if no params found", () => {
+  const url = "";
+
+  expect(fetchQueryParams(url)).toEqual({});
+});
+
+test("fetchQueryParams returns object containing query params", () => {
+  const url = "stringParam=123&arrayParam=1&arrayParam=2";
+
+  expect(fetchQueryParams(url)).toEqual({
+    stringParam: "123",
+    arrayParam: ["1", "2"],
+  });
 });
